UUID to ID Debug Branch (#12)
* Fixed worng id check in display questionary

* Fixed reports and other checks UUID => ID

* gitignore
