+++ QuestionariesUserCases.md
@@ -1,0 +1,38 @@
+# Questionary User Cases 
+
+This is to pass to notion as soon as I get wifi.
+
+
+OPOSICIONES:
+- The administrator can create / edit and delete an opposition with name and date.
+- The administrator can see the list of active oppositions.
+- The administrator can create / edit / delete a topic, selecting the topic group (3 mayor topic groups)  and the name.
+- The administrator can create / edit / delete a subtopic, for a specific topic and selecting a name.
+- The administrator can see all the topics/subtopics link to a opposition
+- For a given topic the administrator can see the list of subtopics, oppositions and questions related to it.
+- The administrator can create / clone / edit / delete and a question.
+- The questions type test they have a question, 4 possible answers 1 is right, explanation and image.
+- The questions type memory card they have a question, 1 answer, explanation and image.
+- The adminsitrator can see the list of questions for a topic / subtopic, the total of questions for each (including subtopic questions) and change the visibility.
+- The administrator can import topics via CSV
+- The administrator can import subtopics via CSV
+- The administrator can import tests and memory card questions via CSV
+- The administrator can see the list of the imports results and the questions imported status of each import.
+- The administrator will get a notification when the import has suceed or not.
+
+
+TESTS:
+- The student can generate a test selecting the type of test (test / memory card) and the number of questions. Selecting the topic group / topics / subtopics that desire to be included.
+- The test will be generating prioritizing the failed questions and answered questions of the student. Adding correct answers to fill up if required.
+- If there is no enough questions, the test will fill up as much questions he has.
+- The student can see the list of uncomplete tests and resume them.
+- The student can reply all the questions of a tests and submit them.
+- The student can see all his memory cards sets and reveal each single card.
+- The student can blame a question, no matter the type of question.
+- A email will be sent to the admin when a question is blamed
+- The student can see all the completed tests and his answer for each of them (wherever was right or wrong)
+- The student can see for each test a summary for each test and the total of questions unanswered, correct or wrong.
+- The student can check by topic which questions he answered wrong (no matter in which test) and the right answer
+- The student can see a statistic of total questions (correct, wrong answered) by topic in the last month, last 3 month, and the total summary.
+- The student can see the 5 top topics that he need to improve off, with the total questions wrong, not answered (and the percentage) with an statistic view.
+
