+++ src/modules/tests/components/Questionnaire/ItemQuestionnaire.vue
@@ -1,17 +1,12 @@
 <template>
   <div>
-    <!-- <p class="text--darken-3 mt-2 mb-3 font-weight-bold"> {{ getIndexQuestion.index }}. - <span class="font-weight-bold">{{ questionWithAnswers?.attributes['question-text'] }}</span></p> -->
-
     <v-list-item-content class="mb-2">
-      <v-list-item-title class="text--darken-3 font-weight-bold">{{ getIndexQuestion.index }}. - {{ questionWithAnswers?.attributes['question-text'] }}</v-list-item-title>
+      <v-list-item-title class="text--darken-3 font-weight-bold">
+        {{ getIndexQuestion.index }}. -
+        {{ questionWithAnswers?.attributes['question-text'] }}
+      </v-list-item-title>
     </v-list-item-content>
-    <!-- <p class="text--darken-3 my-3 font-weight-bold">{{ getIndexQuestion.index }}. - <span class="font-weight-bold">{{ questionWithAnswers?.attributes['question-text'] }}</span></p> -->
-    <!-- <answers-radio-buttons-questionnaire
-      :ref="`answers-group-question-${questionWithAnswers.id}`"
-      :answers="getAnswersOfQuestion"
-      :question-uuid="questionWithAnswers.id"
-      @changeSelectedAnswer="changeSelectionAnswerResolvedInQuestion"
-    /> -->
+
     <answers-list-solve-action-questionnaire
       :ref="`answers-group-question-${questionWithAnswers.id}`"
       :answers="getAnswersOfQuestion"
@@ -34,7 +29,6 @@
 </template>
 
 <script>
-//import AnswersRadioButtonsQuestionnaire from './AnswersRadioButtonsQuestionnaire'
 import AnswersListSolveActionQuestionnaire from './AnswersListSolveActionQuestionnaire'
 import { mapState, mapMutations, mapActions } from 'vuex'
 import SetHistoryAnswersResolved from '../../views/FetchTest/SetHistoryAnswersResolved'
@@ -44,7 +38,6 @@ import _ from 'lodash'
 export default {
   name: 'ItemQuestionnaire',
   components: {
-    //AnswersRadioButtonsQuestionnaire
     AnswersListSolveActionQuestionnaire
   },
   mixins: [SetHistoryAnswersResolved, SaveHistoryAnswersResolved],
@@ -62,19 +55,25 @@ export default {
       required: true
     }
   },
-  data () {
+  data() {
     return {
       answer_selected_id: null
     }
   },
   computed: {
-    ...mapState('testsService', ['questionsDataHistoryByTest', 'questionsDataResolved']),
+    ...mapState('testsService', [
+      'questionsDataHistoryByTest',
+      'questionsDataResolved'
+    ]),
     getIndexQuestion() {
-      return this.questionsDataHistoryByTest.find((question) => question.question_id === this.questionWithAnswers.id)
+      return this.questionsDataHistoryByTest.find(
+        (question) => question.question.uuid === this.questionWithAnswers.id
+      )
     },
     getAnswersOfQuestion() {
-
-      const answers = this.shuffleAnswers(this.questionWithAnswers.relationships['answers-test'].data)
+      const answers = this.shuffleAnswers(
+        this.questionWithAnswers.relationships['answers-test'].data
+      )
 
       return answers.map((answer) => {
         return {
@@ -89,14 +88,17 @@ export default {
       //this.saveAnswerResolvedOfQuestion()
     },
     questionWithAnswers: {
-      handler () {
+      handler() {
         this.setAnswerResolved()
       },
       deep: true
     }
   },
   created() {
-    this.saveAnswerResolvedOfQuestion = _.debounce(this.saveAnswerResolvedOfQuestion, 300)
+    this.saveAnswerResolvedOfQuestion = _.debounce(
+      this.saveAnswerResolvedOfQuestion,
+      300
+    )
   },
   mounted() {
     this.setAnswerResolved()
@@ -104,39 +106,35 @@ export default {
   methods: {
     ...mapMutations('testsService', ['SET_QUESTIONS_DATA_RESOLVED']),
     saveQuestionAnswerInStore() {
-
       const questionsResolved = this.questionsDataResolved
 
-      const questionIndex = this.getResolvedQuestion()
+      const questionIndex = this.questionsDataResolved.findIndex(
+        (question) => question.question.uuid === this.questionUuid
+      )
 
-      if (questionIndex === -1) {
-        questionsResolved.push({
-          question_text: this.questionWithAnswers.attributes.question,
-          answer_text: this.questionWithAnswers.relationships['answers-test'].data.find((answer) => answer.id === this.answer_selected_id),
-          question_id: this.questionWithAnswers.id,
-          answer_id: this.answer_selected_id
-        })
-        this.SET_QUESTIONS_DATA_RESOLVED(questionsResolved)
-
+      if (questionIndex >= 0) {
         return
       }
 
-      questionsResolved[questionIndex]['answer_id'] = this.answer_selected_id
-
-      //console.log(questionsResolved)
+      questionsResolved.push({
+        question_text: this.questionWithAnswers.attributes.question,
+        answer_text: this.questionWithAnswers.relationships[
+          'answers-test'
+        ].data.find((answer) => answer.id === this.answer_selected_id),
+        question_id: this.questionWithAnswers.id,
+        answer_id: this.answer_selected_id
+      })
       this.SET_QUESTIONS_DATA_RESOLVED(questionsResolved)
-    },
-    getResolvedQuestion() {
-      return this.questionsDataResolved.findIndex((question) => question.question_id === this.questionUuid)
     }
   }
 }
 </script>
 
 <style scoped>
-  .v-list-item__subtitle, .v-list-item__title{
-      text-overflow: initial!important;
-      white-space: initial!important;
-      color: #000;
-  }
+.v-list-item__subtitle,
+.v-list-item__title {
+  text-overflow: initial !important;
+  white-space: initial !important;
+  color: #000;
+}
 </style>
